Large refactoring of pricesummary stuff (wip)
